move manager perfmap
